Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 16ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240118/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 long long n, a, b;
 
-long long f(int num) {
+long long f(long long num) {
 	long long left = 1, right = n;
 	long long cnt = 1;
 	long long mid = 0;
